refac: use double precision for raw material quantities + add method to reset product's global id correctly

+++ src/main/java/domain/entities/product/KitCopoDeVidro.java
@@ -2,7 +2,7 @@ package domain.entities.product;
 
 public class KitCopoDeVidro extends Product {
 
-    public KitCopoDeVidro(String name, int rawMaterialPerUnit) {
+    public KitCopoDeVidro(String name, double rawMaterialPerUnit) {
         super(name, rawMaterialPerUnit, 1.0, 0.0);
     }
